Implementando Spring Security V3.0

+++ src/main/java/com/optilife/security/TokenServiceImpl.java
@@ -2,18 +2,12 @@ package com.optilife.security;
 
 import com.optilife.model.entity.Token;
 import com.optilife.model.entity.Usuario;
-import io.jsonwebtoken.Jwts;
-import io.jsonwebtoken.SignatureAlgorithm;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Service;
 
-import java.util.Date;
-import java.time.Instant;
 import java.time.LocalDateTime;
-import java.time.temporal.ChronoUnit;
 import java.util.Optional;
-import java.util.UUID;
+
 
 @Service
 public class TokenServiceImpl implements TokenService {
@@ -21,33 +15,30 @@ public class TokenServiceImpl implements TokenService {
     @Autowired
     private TokenRepository tokenRepository;
 
-    @Value("${jwt.secret}")
-    private String jwtSecret;
+    @Autowired
+    private TokenProvider tokenProvider;
 
     @Override
     public String generarToken(Usuario usuario) {
-        //generator UUID unico para usuario
-        String idUsuario = UUID.randomUUID().toString();
-        //generator token JWT
-        String token = Jwts.builder()
-                .setSubject(idUsuario)
-                .setIssuedAt(new Date())
-                .setExpiration(Date.from(Instant.now().plus(24, ChronoUnit.HOURS)))
-                .signWith(SignatureAlgorithm.HS512, jwtSecret)
-                .compact();
-        //creando nuevo token con el JWT y fecha de expiración
-        Token nuevoToken = new Token(token,usuario,LocalDateTime.now().plusHours(24));
-        tokenRepository.save(nuevoToken);//gurdando en DB
+        //generando token JWT a través de TokenProvider
+        String token = tokenProvider.generarToken(usuario);
+
+        //crear y guardar token en la DB con expiración
+        Token newToken = new Token(token, usuario, LocalDateTime.now().plusHours(24));
+        tokenRepository.save(newToken);
         return token;
     }
 
+    @Override
     public Optional<Usuario> validarToken(String token) {
         Optional<Token> optionalToken = tokenRepository.findByToken(token);
-        if (optionalToken.isPresent()) {
+        if (optionalToken.isPresent() && tokenProvider.validarToken(token)) {
             Token tokenEntity = optionalToken.get();
-            if(tokenEntity.getFechaExpiracion().isAfter(LocalDateTime.now())) {
+
+            //verificando si el token expiró o no
+            if (tokenEntity.getFechaExpiracion().isAfter(LocalDateTime.now())) {
                 return Optional.of(tokenEntity.getUsuario());
-            }else {
+            } else {
                 tokenRepository.delete(tokenEntity);
             }
         }
